audio: silence some int/float conversion/narrowing warnings
Pick-to: 6.12

+++ src/multimedia/audio/qaudiosystem_platform_stream_support.cpp
@@ -439,7 +439,7 @@ std::chrono::microseconds QPlatformAudioSourceStream::processedDuration() const
 {
     return std::chrono::microseconds{
         m_format.durationForFrames(
-                m_totalNumberOfFramesPushedToRingbuffer.load(std::memory_order_relaxed)),
+                qint32(m_totalNumberOfFramesPushedToRingbuffer.load(std::memory_order_relaxed))),
     };
 }
